+++ lib/entities/android_params.g.dart
@@ -11,14 +11,18 @@ AndroidParams _$AndroidParamsFromJson(Map<String, dynamic> json) =>
       isCustomNotification: json['isCustomNotification'] as bool?,
       isCustomSmallExNotification: json['isCustomSmallExNotification'] as bool?,
       isShowLogo: json['isShowLogo'] as bool?,
+      isShowCallID: json['isShowCallID'] as bool?,
       ringtonePath: json['ringtonePath'] as String?,
       backgroundColor: json['backgroundColor'] as String?,
       backgroundUrl: json['backgroundUrl'] as String?,
       actionColor: json['actionColor'] as String?,
+      textColor: json['textColor'] as String?,
       incomingCallNotificationChannelName:
           json['incomingCallNotificationChannelName'] as String?,
       missedCallNotificationChannelName:
           json['missedCallNotificationChannelName'] as String?,
+      isShowFullLockedScreen:
+      json['isShowFullLockedScreen'] as bool?,
     );
 
 Map<String, dynamic> _$AndroidParamsToJson(AndroidParams instance) =>
@@ -26,12 +30,15 @@ Map<String, dynamic> _$AndroidParamsToJson(AndroidParams instance) =>
       'isCustomNotification': instance.isCustomNotification,
       'isCustomSmallExNotification': instance.isCustomSmallExNotification,
       'isShowLogo': instance.isShowLogo,
+      'isShowCallID': instance.isShowCallID,
       'ringtonePath': instance.ringtonePath,
       'backgroundColor': instance.backgroundColor,
       'backgroundUrl': instance.backgroundUrl,
       'actionColor': instance.actionColor,
+      'textColor': instance.textColor,
       'incomingCallNotificationChannelName':
           instance.incomingCallNotificationChannelName,
       'missedCallNotificationChannelName':
           instance.missedCallNotificationChannelName,
+      'isShowFullLockedScreen': instance.isShowFullLockedScreen,
     };