+++ lib/flutter_callkit_incoming.dart
@@ -49,6 +49,12 @@ class FlutterCallkitIncoming {
     await _channel.invokeMethod("showMissCallNotification", params.toJson());
   }
 
+  /// Hide notification call for Android.
+  /// Only Android
+  static Future hideCallkitIncoming(CallKitParams params) async {
+    await _channel.invokeMethod("hideCallkitIncoming", params.toJson());
+  }
+
   /// Start an Outgoing call.
   /// On iOS, using Callkit(create a history into the Phone app).
   /// On Android, Nothing(only callback event listener).
@@ -109,6 +115,16 @@ class FlutterCallkitIncoming {
   /// On Android: return Empty
   static Future getDevicePushTokenVoIP() async {
     return await _channel.invokeMethod("getDevicePushTokenVoIP");
+  }
+
+  /// Silence CallKit events
+  static Future silenceEvents() async {
+    return await _channel.invokeMethod("silenceEvents", true);
+  }
+
+  /// Unsilence CallKit events
+  static Future unsilenceEvents() async {
+    return await _channel.invokeMethod("silenceEvents", false);
   }
 
   /// Request permisstion show notification for Android(13)